requirements.txt
discord.py[voice]==1.6
sqlalchemy==1.3.23
psycopg2-binary==2.8.6
requests==2.25.1
pytz==2021.1
inflect==5.3.0
aiohttp==3.7.4
beautifulsoup4==4.9.3
aiolimiter==1.0.0b1
